Aula 11/09

+++ src/main/kotlin/aula0909/exercicios/santander/Cliente.kt
@@ -1,3 +1,3 @@
 package aula0909.exercicios.santander
 
-class Cliente(val numero: Int, val sobrenome: String, val rg: String, val cpf: Int)
+class Cliente(val numero: Int, val sobrenome: String, val rg: String, val cpf: Long)
